Support more layouts and checksum fixing
Welcome to the worst part of P4. TCP option parsing with unions and
vector still does not work. While the recomputation of TCP checksum
(with option) is not even speaken of...

So we fix this in P4_16 (other platforms have other externs) by parsing
a few known layouts (Linux for now) and doing an incremental checksum
fixing.